+++ src/modules/support/classes/Ticket.ts
@@ -6,14 +6,15 @@ import channels from "../../../channels";
 import { pmdDB } from "../../../database/client";
 
 const coll = pmdDB.collection("tickets"),
-	circleFolder =
-		"https://raw.githubusercontent.com/PreMiD/Discord-Bot/master/.discord/";
+	circleFolder = "https://github.com/PreMiD/Discord-Bot/blob/main/.discord/";
 
 let ticketCount = 0;
 
 export class Ticket {
 	id: string;
 	status: number;
+	ticketContent: string;
+	attachments: Array<string>;
 
 	ticketMessage: Discord.Message;
 	user: Discord.GuildMember;
@@ -26,28 +27,25 @@ export class Ticket {
 	embed: Discord.MessageEmbedOptions;
 
 	attachmentsMessage: Discord.Message;
-
 	constructor() {}
 
-	async fetch(type: "ticket" | "message" | "channel", arg: any) {
+	async fetch(type: "ticket" | "message" | "channel" | "author", arg: any) {
+
 		const ticket =
 			type === "ticket"
 				? arg
-				: await coll.findOne(
-						type === "message"
-							? { ticketMessage: arg }
-							: { supportChannel: arg }
-				  );
+				: type === "author"
+				? await coll.findOne({ userId: arg })
+				: await coll.findOne(type === "message" ? { ticketMessage: arg } : { supportChannel: arg });
 
 		if (!ticket) return false;
 
 		this.id = ticket.ticketId;
 		this.status = ticket.status;
+		this.attachments = ticket.attachments;
 
 		try {
-			this.ticketMessage = await (client.guilds.cache
-				.first()
-				.channels.cache.get(
+			this.ticketMessage = await ((client.channels.cache.get(channels.ticketCategory) as Discord.TextChannel).guild.channels.cache.get(
 					channels.ticketChannel
 				) as Discord.TextChannel).messages.fetch(ticket.ticketMessage);
 			this.embed = this.ticketMessage.embeds[0];
@@ -56,30 +54,26 @@ export class Ticket {
 		}
 
 		if (this.status === 1) {
-			this.channel = client.guilds.cache
-				.first()
-				.channels.cache.get(ticket.supportChannel) as Discord.TextChannel;
+			this.channel = (client.channels.cache.get(channels.ticketCategory) as Discord.TextChannel).guild.channels.cache
+				.get(ticket.supportChannel) as Discord.TextChannel;
 			this.channelMessage = await this.channel?.messages.fetch(
 				ticket.supportEmbed
 			);
 			this.supporters = await Promise.all(
 				ticket.supporters.map((s: string) =>
-					client.guilds.cache.first().members.fetch(s)
+					(client.channels.cache.get(channels.ticketCategory) as Discord.TextChannel).guild.members.fetch(s)
 				)
 			);
 		}
 
 		if (ticket.attachmentMessage)
-			this.attachmentsMessage = await (client.guilds.cache
-				.first()
-				.channels.cache.get(
+			this.attachmentsMessage = await ((client.channels.cache.get(channels.ticketCategory) as Discord.TextChannel).guild.channels.cache
+				.get(
 					channels.ticketChannel
 				) as Discord.TextChannel).messages.fetch(ticket.attachmentMessage);
 
 		try {
-			this.user = await client.guilds.cache
-				.first()
-				.members.fetch(ticket.userId);
+			this.user = await (client.channels.cache.get(channels.ticketCategory) as Discord.TextChannel).guild.members.fetch(ticket.userId);
 		} catch (_) {}
 		return true;
 	}
@@ -92,10 +86,14 @@ export class Ticket {
 
 			this.id = ticketCount.toString().padStart(5, "0");
 
+			this.ticketContent = message.cleanContent;
+
+			this.attachments = [];
+
 			this.embed = {
 				author: {
 					name: `Ticket#${this.id} [OPEN]`,
-					iconURL: `${circleFolder}green_circle.png`
+					iconURL: `${circleFolder}green_circle.png?raw=true`
 				},
 				description: message.cleanContent,
 				footer: {
@@ -104,6 +102,15 @@ export class Ticket {
 				},
 				color: "#77ff77"
 			};
+
+			if (message.attachments.size > 0) {
+				this.attachments.push(`[${message.attachments.first().name}](${message.attachments.first().proxyURL})`)
+				this.embed.fields = [{
+					name: "Attachments",
+					value: this.attachments.join(", "),
+					inline: false
+				}]
+			}
 
 			this.ticketMessage = await (message.guild.channels.cache.get(
 				channels.ticketChannel
@@ -115,21 +122,12 @@ export class Ticket {
 				.react("🚫")
 				.then(() =>
 					this.ticketMessage.react(
-						message.guild.emojis.cache.get("521018476870107156")
+						client.guilds.cache.get("493130730549805057").emojis.cache.get("521018476870107156")
 					)
 				);
 
-			if (message.attachments.size > 0)
-				this.attachmentsMessage = await (client.guilds.cache
-					.first()
-					.channels.cache.get(
-						channels.ticketChannel
-					) as Discord.TextChannel).send(message.attachments.first());
-
 			message.author
-				.send(
-					`Your ticket \`\`#${this.id}\`\` has been submitted and will be answered shortly. Please be patient. Thank you!`
-				)
+				.send(`Your ticket \`\`#${this.id}\`\` has been submitted and will be answered shortly.`)
 				.catch(() => {});
 
 			coll.insertOne({
@@ -137,15 +135,16 @@ export class Ticket {
 				userId: message.author.id,
 				ticketMessage: this.ticketMessage.id,
 				timestamp: Date.now(),
-				attachmentMessage: this.attachmentsMessage
-					? this.attachmentsMessage.id
-					: undefined,
+				attachments: this.attachments,
 				created: Date.now()
 			});
 
 			message.delete().catch(() => {});
+			(client.channels.cache.get(channels.supportChannel) as Discord.TextChannel).updateOverwrite(message.author.id, {
+				SEND_MESSAGES: false
+			})
 		} catch (err) {
-			(message.guild.channels.cache.get(
+			(client.channels.cache.get(
 				channels.dev
 			) as Discord.TextChannel).send(
 				new Discord.MessageEmbed({
@@ -158,17 +157,11 @@ export class Ticket {
 
 	async accept(supporter: Discord.GuildMember) {
 		if (
-			(client.guilds.cache
-				.first()
-				.channels.resolve(channels.ticketCategory) as Discord.CategoryChannel)
+			(client.channels.resolve(channels.ticketCategory) as Discord.CategoryChannel)
 				.children.size >= 50
 		) {
 			(
-				await (client.guilds.cache
-					.first()
-					.channels.resolve(
-						channels.ticketChannel
-					) as Discord.TextChannel).send(
+				await (client.channels.resolve(channels.ticketChannel) as Discord.TextChannel).send(
 					`${supporter.toString()}, Can't accept ticket, the category limit has been reached.`
 				)
 			).delete({ timeout: 15 * 1000 });
@@ -180,8 +173,7 @@ export class Ticket {
 
 		this.embed.author = {
 			name: `Ticket#${this.id} [PENDING]`,
-			iconURL:
-				"https://raw.githubusercontent.com/PreMiD/Discord-Bot/master/.discord/yellow_circle.png"
+			iconURL: "https://github.com/PreMiD/Discord-Bot/blob/main/.discord/yellow_circle.png?raw=true"
 		};
 		this.embed.color = "#f4dd1a";
 
@@ -197,13 +189,14 @@ export class Ticket {
 			"USE_EXTERNAL_EMOJIS"
 		];
 
-		this.channel = (await client.guilds.cache.first().channels.create(this.id, {
+		this.channel = (
+			await (client.channels.cache.get(channels.ticketCategory) as Discord.CategoryChannel).guild.channels.create(this.id, {
 			parent: channels.ticketCategory,
 			type: "text",
 			//@ts-ignore
 			permissionOverwrites: [
 				{
-					id: client.guilds.cache.first().id,
+					id: supporter.guild.id,
 					deny: ["VIEW_CHANNEL"]
 				},
 				{
@@ -241,10 +234,17 @@ export class Ticket {
 				inline: true
 			}
 		];
+
+		if(this.attachments.length > 0) this.embed.fields["push"]({
+			name: "Attachments", 
+			value: this.attachments.join(", "),
+			inline: true
+		})
+		
 		this.ticketMessage.edit(this.embed);
 
 		//@ts-ignore False types...
-		this.embed.fields.pop();
+		this.embed.fields = this.embed.fields.filter(x => x.name != "Channel");
 		this.embed.footer = {
 			text: "p!close - Closes this ticket."
 		};
@@ -270,31 +270,64 @@ export class Ticket {
 		sortTickets();
 	}
 
-	async close(closer?: Discord.GuildMember, reason?: string) {
-		if (reason)
-			this.user
-				.send(
-					`Your Ticket \`\`#${this.id}\`\` has been closed. Reason:\n\n*\`\`${reason}\`\`*`
-				)
-				.catch(() => {});
+	async close(closer?: any, reason?: string) {
+
+		this.user.send(`Your ticket \`\`#${this.id}\`\` has been closed by **${closer.tag ? closer.tag : closer.user.tag}**. Reason: \`\`${reason || "Not Specified"}\`\``).catch(() => {});
+		
+		const getVars = url => 
+		    /^https:\/\/discordapp\.com\/api\/webhooks\/(\d{18})\/([\w-]{1,})$/.test(url) ? {
+		    id: /^https:\/\/discordapp\.com\/api\/webhooks\/(\d{18})\/([\w-]{1,})$/.exec(url)[1],
+		    token: /^https:\/\/discordapp\.com\/api\/webhooks\/(\d{18})\/([\w-]{1,})$/.exec(url)[2]
+		}
+		: null,
+		vars = getVars(process.env.ticketLogsWebhook),
+		webhook = new Discord.WebhookClient(vars.id, vars.token),
+		embed = new Discord.MessageEmbed()
+			.setAuthor(`Ticket#${this.id} [CLOSED]`, "https://github.com/PreMiD/Discord-Bot/blob/main/.discord/red_circle.png?raw=true")
+			.setColor("#b52222")
+			.setDescription(this.embed.description)
+			.addFields([
+				{
+					name: `Opened By`,
+					value: this.user.user.tag,
+					inline: true
+				},
+				{
+					name: `Closed By`,
+					value: closer.tag ? closer.tag : closer.user.tag,
+					inline: true
+				},
+				{
+					name: `Reason`,
+					value: reason || "Not Specified",
+					inline: true
+				},
+				{
+					name: `Supporter(s)`,
+					value: this.supporters,
+					inline: true
+				},
+				{
+					name: "Attachments",
+					value: this.attachments.length > 0 ? this.attachments.join(", ") : "None",
+					inline: true
+				}
+			]);
+
+		webhook.send(embed);
 
 		if (this.embed.thumbnail) delete this.embed.thumbnail;
 		delete this.embed.fields;
-
-		if (this.attachmentsMessage && this.attachmentsMessage.deletable)
-			this.attachmentsMessage.delete();
-
+		if (this.attachmentsMessage && this.attachmentsMessage.deletable) this.attachmentsMessage.delete();
 		if (this.ticketMessage.deletable) this.ticketMessage.delete();
-		if (this.channel.deletable) this.channel.delete();
+	    if (this.channel && this.channel.deletable) this.channel.delete();
+		
+		(client.channels.cache.get(channels.supportChannel) as Discord.TextChannel).permissionOverwrites.get(this.user.id).delete()
 
 		coll.findOneAndUpdate(
-			{ supportChannel: this.channel.id },
-			{
+			{ supportChannel: this.channel ? this.channel.id : 0}, {
 				$unset: { supportChannel: "", supportEmbed: "" },
-				$set: {
-					status: 2,
-					closer: closer?.id || undefined
-				}
+				$set: { status: 2, closer: closer.id	}
 			}
 		);
 	}
@@ -307,7 +340,7 @@ export class Ticket {
 		//@ts-ignore False types...
 		this.embed.fields[0] = {
 			name: "Supporter",
-			value: this.supporters.join(", ")
+			value: this.supporters.join(", ") || "None"
 		};
 
 		this.ticketMessage.edit(this.embed);
@@ -376,6 +409,35 @@ export class Ticket {
 				{ $set: { supporters: this.supporters.map(s => s.id) } }
 			);
 		}
+	}
+
+	async attach(imageObj, userId) {
+		const { attachments } = await coll.findOne({userId: userId, status: 1});
+		attachments.push(`[${imageObj.name}](${imageObj.proxyURL})`);
+		this.embed.fields ?
+			this.embed.fields.filter(x => x.name == "Attachments").length == 1 ?
+				(this.embed.fields.filter(x => x.name == "Attachments")[0].value = attachments.join(", "))
+			: this.embed.fields['push']({name: "Attachments", value: attachments.join(", "), inline: true})
+		: this.embed.fields = [{name: "Attachments", value: attachments.join(", "), inline: true}];
+		
+		this.channelMessage.edit(this.embed);
+
+		let emb = this.ticketMessage.embeds[0];
+
+		emb.fields ?
+			emb.fields.filter(x => x.name == "Attachments").length == 1 ?
+				(emb.fields.filter(x => x.name == "Attachments")[0].value = attachments.join(", "))
+			: emb.fields["push"]({name: "Attachments", value: attachments.join(", "), inline: true})
+		: emb.fields = [{name: "Attachments", value: attachments.join(", "), inline: true}];
+
+		this.ticketMessage.edit(emb);
+
+		coll.findOneAndUpdate({userId: userId}, { 
+			$set: {
+				attachments: attachments
+			}
+		})
+
 	}
 
 	async sendCloseWarning() {
